requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-groq==0.1.2
python-dotenv==1.0.0
requests==2.31.0